requirements.txt
alembic==1.13.2
fastapi==0.111.0
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.34
uvicorn[standard]==0.30.5
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-multipart==0.0.9
pymysql==1.1.2
passlib[argon2]